pyproject.toml
[tool.poetry]
name = "OllaBot"
version = "0.1.0"
description = ""
authors = ["Jacob Celestine"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.12.17"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.38"
ollama = "^0.4.7"
llama-index-llms-ollama = "^0.5.2"
llama-index-embeddings-ollama = "^0.5.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
streamlit = "^1.42.0"
mypy = "^1.15.0"
ruff = "^0.9.6"
types-pyyaml = "^6.0.12.20241230"
watchdog = "^6.0.0"
fastapi = "^0.115.8"
types-requests = "^2.32.0.20241016"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 200


[tool.ruff.lint]
select = [ "E", "F", "Q", "UP", "B", "S", "RUF", "C", "ERA", "PIE", "SIM", "A", "G",]
ignore = [ "ERA001", "C901", "SIM112"]


[tool.ruff.lint.per-file-ignores]
"tests/*" = [ "D", "S",]


[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
no_implicit_optional = true
disallow_untyped_defs = true
namespace_packages = true
explicit_package_bases = true


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
